Tidy up the whole modules
