+++ src/array/checkArray.js
@@ -1,50 +1,19 @@
 /* eslint-disable implicit-arrow-linebreak */
 const checkArray = (list, flag) => {
-  let greater = 0;
-  let less = 0;
-  let equal = 0;
-
-  list.forEach((item) => {
-    if (item < flag) {
-      less += 1;
-    } else if (item > flag) {
-      greater += 1;
-    } else {
-      equal += 1;
-    }
-  });
-
-  return { greater, less, equal };
+  return 'not implemented yet';
 };
 
-const isAlbertHere = array =>
-  array.some(item => item === 'albert');
+const isAlbertHere = array => 'not implemented yet';
 
-const findAllBoys = (array) => {
-  const boyArray = [];
-  array.forEach((value) => {
-    if (value.gender === 'M'){
-      boyArray.push(value.name);
-    }
-  });
-  return boyArray;
+const findAllBoys = array => {
+  return 'not implemented yet';
 };
 
-const copyArray = array => [...array];
+const copyArray = array => 'not implemented yet';
 
-const countBoysAndGirls = (array) => {
-  let boys = 0;
-  let girls = 0;
-  array.forEach((value) => {
-    if (value.gender === 'M') {
-      boys += 1;
-    }
-    else if (value.gender === 'F') {
-      girls += 1;
-    }
-  });
-  return { boys, girls };
-}
+const countBoysAndGirls = array => {
+  return 'not implemented yet';
+};
 
 exports.checkArray = checkArray;
 exports.isAlbertHere = isAlbertHere;
