+++ src/compute.js
@@ -1,0 +1,8 @@
+//  check value is not number return true
+//  if value is number return false
+
+function compute() {
+  return 'not implemented yet';
+}
+
+exports.compute = compute;
