+++ src/isEven.js
@@ -1,12 +1,5 @@
 /* eslint-disable no-confusing-arrow,implicit-arrow-linebreak */
-const getTypeName = value =>
-  value == null
-    ? ''
-    : ` (${typeof value})`;
 
-const isEven = n =>
-  typeof n === 'number'
-    ? n % 2 === 0
-    : `expect number but received ${n}${getTypeName(n)}`;
+const isEven = n => 'not implemented yet';
 
 exports.isEven = isEven;
